// cluster/portal/portal/web/package/src/apis/userv1/userv1.ts
export interface DNS {
    /**
     * Servers is the list of the Cluster private DNS servers.
     *
     * @generated from protobuf field: repeated string servers = 1;
     */
    servers: string[];
}
// ...
export interface Service {
    /**
     * @generated from protobuf field: string apiVersion = 1;
     */
    apiVersion: string;
    /**
     * @generated from protobuf field: string kind = 2;
     */
    kind: string;
    /**
     * @generated from protobuf field: octelium.api.main.meta.v1.Metadata metadata = 3;
     */
    metadata?: Metadata;
    /**
     * @generated from protobuf field: octelium.api.main.user.v1.Service.Spec spec = 4;
     */
    spec?: Service_Spec;
    /**
     * @generated from protobuf field: octelium.api.main.user.v1.Service.Status status = 5;
     */
    status?: Service_Status;
}
/**
 * @generated from protobuf message octelium.api.main.user.v1.Service.Spec
 */
export interface Service_Spec {
    /**
     * Port is the Service listen port
     *
     * @generated from protobuf field: uint32 port = 1;
     */
    port: number;
    /**
     * Type is the Service tpye
     *
     * @generated from protobuf field: octelium.api.main.user.v1.Service.Spec.Type type = 2;
     */
    type: Service_Spec_Type;
    /**
     * IsTLS shows whether the Service is listening over TLS
     *
     * @generated from protobuf field: bool isTLS = 3;
     */
    isTLS: boolean;
    /**
     * IsPublic means that the Service is publicly exposed and can be accessed
     * via the client-less/BeyondCorp mode
     *
     * @generated from protobuf field: bool isPublic = 4;
     */
    isPublic: boolean;
}
/**
 * @generated from protobuf enum octelium.api.main.user.v1.Service.Spec.Type
 */
export enum Service_Spec_Type {
    /**
     * @generated from protobuf enum value: UNSET = 0;
     */
    UNSET = 0,
    /**
     * @generated from protobuf enum value: TCP = 1;
     */
    TCP = 1,
    /**
     * @generated from protobuf enum value: UDP = 2;
     */
    UDP = 2,
    /**
     * @generated from protobuf enum value: HTTP = 3;
     */
    HTTP = 3,
    /**
     * @generated from protobuf enum value: SSH = 4;
     */
    SSH = 4,
    /**
     * @generated from protobuf enum value: KUBERNETES = 5;
     */
    KUBERNETES = 5,
    /**
     * @generated from protobuf enum value: POSTGRES = 6;
     */
    POSTGRES = 6,
    /**
     * @generated from protobuf enum value: MYSQL = 7;
     */
    MYSQL = 7,
    /**
     * @generated from protobuf enum value: GRPC = 8;
     */
    GRPC = 8,
    /**
     * @generated from protobuf enum value: WEB = 9;
     */
    WEB = 9,
    /**
     * @generated from protobuf enum value: DNS = 10;
     */
    DNS = 10
}
/**
 * @generated from protobuf message octelium.api.main.user.v1.Service.Status
 */
export interface Service_Status {
    /**
     * Namespace is the Service's Namespace
     *
     * @generated from protobuf field: string namespace = 1;
     */
    namespace: string;
    /**
     * Addresses is the list of private addresses used by the Service
     *
     * @generated from protobuf field: repeated string addresses = 2;
     */
    addresses: string[];
    /**
     * PrimaryHostname is the unique primary hostname
     *
     * @generated from protobuf field: string primaryHostname = 3;
     */
    primaryHostname: string;
}
/**
 * @generated from protobuf message octelium.api.main.user.v1.ServiceList
 */
export interface ServiceList {
    /**
     * @generated from protobuf field: string apiVersion = 1;
     */
    apiVersion: string;
    /**
     * @generated from protobuf field: string kind = 2;
     */
    kind: string;
    /**
     * @generated from protobuf field: repeated octelium.api.main.user.v1.Service items = 3;
     */
    items: Service[];
    /**
     * ListResponseMeta is common information about the list.
     *
     * @generated from protobuf field: octelium.api.main.meta.v1.ListResponseMeta listResponseMeta = 4;
     */
    listResponseMeta?: ListResponseMeta;
}
/**
 * @generated from protobuf message octelium.api.main.user.v1.ListServiceOptions
 */
export interface ListServiceOptions {
    /**
     * @generated from protobuf field: octelium.api.main.meta.v1.CommonListOptions common = 1;
     */
    common?: CommonListOptions;
    /**
     * @generated from protobuf field: string namespace = 2;
     */
    namespace: string;
}
/**
 * @generated from protobuf message octelium.api.main.user.v1.ListNamespaceOptions
 */
export interface ListNamespaceOptions {
    /**
     * @generated from protobuf field: octelium.api.main.meta.v1.CommonListOptions common = 1;
     */
    common?: CommonListOptions;
}
// ...
class DNS$Type extends MessageType<DNS> {
    constructor() {
        super("octelium.api.main.user.v1.DNS", [
            { no: 1, name: "servers", kind: "scalar", repeat: 2 /*RepeatType.UNPACKED*/, T: 9 /*ScalarType.STRING*/ }
        ]);
    }
    create(value?: PartialMessage<DNS>): DNS {
        const message = globalThis.Object.create((this.messagePrototype!));
        message.servers = [];
        if (value !== undefined)
            reflectionMergePartial<DNS>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: DNS): DNS {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* repeated string servers */ 1:
                    message.servers.push(reader.string());
                    break;
                default:
                    let u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
                    let d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
            }
        }
        return message;
    }
    internalBinaryWrite(message: DNS, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        /* repeated string servers = 1; */
        for (let i = 0; i < message.servers.length; i++)
            writer.tag(1, WireType.LengthDelimited).string(message.servers[i]);
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message octelium.api.main.user.v1.DNS
 */
export const DNS = new DNS$Type();
// ...
export const Service = new Service$Type();
// ...
export const Service_Spec = new Service_Spec$Type();
// ...
export const Service_Status = new Service_Status$Type();
// ...
export const ServiceList = new ServiceList$Type();
// @generated message type with reflection information, may provide speed optimized methods
class ListServiceOptions$Type extends MessageType<ListServiceOptions> {
    constructor() {
        super("octelium.api.main.user.v1.ListServiceOptions", [
            { no: 1, name: "common", kind: "message", T: () => CommonListOptions },
            { no: 2, name: "namespace", kind: "scalar", T: 9 /*ScalarType.STRING*/ }
        ]);
    }
    create(value?: PartialMessage<ListServiceOptions>): ListServiceOptions {
        const message = globalThis.Object.create((this.messagePrototype!));
        message.namespace = "";
        if (value !== undefined)
            reflectionMergePartial<ListServiceOptions>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: ListServiceOptions): ListServiceOptions {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* octelium.api.main.meta.v1.CommonListOptions common */ 1:
                    message.common = CommonListOptions.internalBinaryRead(reader, reader.uint32(), options, message.common);
                    break;
                case /* string namespace */ 2:
                    message.namespace = reader.string();
                    break;
                default:
                    let u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
                    let d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
            }
        }
        return message;
    }
    internalBinaryWrite(message: ListServiceOptions, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        /* octelium.api.main.meta.v1.CommonListOptions common = 1; */
        if (message.common)
            CommonListOptions.internalBinaryWrite(message.common, writer.tag(1, WireType.LengthDelimited).fork(), options).join();
        /* string namespace = 2; */
        if (message.namespace !== "")
            writer.tag(2, WireType.LengthDelimited).string(message.namespace);
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message octelium.api.main.user.v1.ListServiceOptions
 */
export const ListServiceOptions = new ListServiceOptions$Type();
// @generated message type with reflection information, may provide speed optimized methods
class ListNamespaceOptions$Type extends MessageType<ListNamespaceOptions> {
    constructor() {
        super("octelium.api.main.user.v1.ListNamespaceOptions", [
            { no: 1, name: "common", kind: "message", T: () => CommonListOptions }
        ]);
    }
    create(value?: PartialMessage<ListNamespaceOptions>): ListNamespaceOptions {
        const message = globalThis.Object.create((this.messagePrototype!));
        if (value !== undefined)
            reflectionMergePartial<ListNamespaceOptions>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: ListNamespaceOptions): ListNamespaceOptions {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* octelium.api.main.meta.v1.CommonListOptions common */ 1:
                    message.common = CommonListOptions.internalBinaryRead(reader, reader.uint32(), options, message.common);
                    break;
                default:
                    let u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
                    let d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
            }
        }
        return message;
    }
    internalBinaryWrite(message: ListNamespaceOptions, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        /* octelium.api.main.meta.v1.CommonListOptions common = 1; */
        if (message.common)
            CommonListOptions.internalBinaryWrite(message.common, writer.tag(1, WireType.LengthDelimited).fork(), options).join();
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message octelium.api.main.user.v1.ListNamespaceOptions
 */
export const ListNamespaceOptions = new ListNamespaceOptions$Type();